Somewhere In the Middle
The structure is....70-80% layed out
I placed almost all the content
Started applying colors
Applied the typefaces/webfonts
Starting to remove 'test' css and html but there is plenty left. Fixed
the animations on the logos to be more midnight sky-light gradient glow
and less blade runner white streak from the middle. All the files from
the past animation are in the img folder and need deleted